clean assets
