--- assets/ml/clap_v1/model_meta.json
@@ -1,7 +1,0 @@
-{
-  "model": "laion/clap-htsat-fused",
-  "input_names": ["audio"],
-  "input_shapes": [[1, 1, 480000]],
-  "opset": 17,
-  "embedding_dim": 512
-}
